# backend/scripts/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Campus Operations AI API",
                 "description":  "API for Campus Operations Intelligence Platform",
                 "version":  "0.0.1"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "security":  [
                     {
                         "bearerAuth":  [

                                        ]
                     }
                 ],
    "paths":  {
...
                  "/api/courses/{id}":  {
                                            "get":  {
                                                        "tags":  [
                                                                     "course-controller"
                                                                 ],
                                                        "operationId":  "getCourseById",
                                                        "parameters":  [
                                                                           {
                                                                               "name":  "id",
                                                                               "in":  "path",
                                                                               "required":  true,
                                                                               "schema":  {
                                                                                              "type":  "integer",
                                                                                              "format":  "int64"
                                                                                          }
                                                                           }
                                                                       ],
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "OK",
                                                                                      "content":  {
                                                                                                      "*/*":  "@{schema=}"
                                                                                                  }
                                                                                  }
                                                                      },
                                                        "security":  [
                                                                         {
                                                                             "bearerAuth":  [

                                                                                            ]
                                                                         }
                                                                     ]
                                                    },
                                            "put":  {
                                                        "tags":  [
                                                                     "course-controller"
                                                                 ],
                                                        "operationId":  "updateCourse",
                                                        "parameters":  [
                                                                           {
                                                                               "name":  "id",
                                                                               "in":  "path",
                                                                               "required":  true,
                                                                               "schema":  {
                                                                                              "type":  "integer",
                                                                                              "format":  "int64"
                                                                                          }
                                                                           }
                                                                       ],
                                                        "requestBody":  {
                                                                            "content":  {
                                                                                            "application/json":  {
                                                                                                                     "schema":  "@{$ref=#/components/schemas/UpdateCourseRequest}"
                                                                                                                 }
                                                                                        },
                                                                            "required":  true
                                                                        },
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "OK",
                                                                                      "content":  {
                                                                                                      "*/*":  "@{schema=}"
                                                                                                  }
                                                                                  }
                                                                      },
                                                        "security":  [
                                                                         {
                                                                             "bearerAuth":  [

                                                                                            ]
                                                                         }
                                                                     ]
                                                    },
                                            "delete":  {
                                                           "tags":  [
                                                                        "course-controller"
                                                                    ],
                                                           "operationId":  "deleteCourse",
                                                           "parameters":  [
                                                                              {
                                                                                  "name":  "id",
                                                                                  "in":  "path",
                                                                                  "required":  true,
                                                                                  "schema":  {
                                                                                                 "type":  "integer",
                                                                                                 "format":  "int64"
                                                                                             }
                                                                              }
                                                                          ],
                                                           "responses":  {
                                                                             "200":  {
                                                                                         "description":  "OK",
                                                                                         "content":  {
                                                                                                         "*/*":  "@{schema=}"
                                                                                                     }
                                                                                     }
                                                                         },
                                                           "security":  [
                                                                            {
                                                                                "bearerAuth":  [

                                                                                               ]
                                                                            }
                                                                        ]
                                                       }
                                        },
...
                  "/api/courses":  {
                                       "get":  {
                                                   "tags":  [
                                                                "course-controller"
                                                            ],
                                                   "operationId":  "getAllCourses",
                                                   "responses":  {
                                                                     "200":  {
                                                                                 "description":  "OK",
                                                                                 "content":  {
                                                                                                 "*/*":  "@{schema=}"
                                                                                             }
                                                                             }
                                                                 },
                                                   "security":  [
                                                                    {
                                                                        "bearerAuth":  [

                                                                                       ]
                                                                    }
                                                                ]
                                               },
                                       "post":  {
                                                    "tags":  [
                                                                 "course-controller"
                                                             ],
                                                    "operationId":  "createCourse",
                                                    "requestBody":  {
                                                                        "content":  {
                                                                                        "application/json":  {
                                                                                                                 "schema":  "@{$ref=#/components/schemas/CreateCourseRequest}"
                                                                                                             }
                                                                                    },
                                                                        "required":  true
                                                                    },
                                                    "responses":  {
                                                                      "200":  {
                                                                                  "description":  "OK",
                                                                                  "content":  {
                                                                                                  "*/*":  "@{schema=}"
                                                                                              }
                                                                              }
                                                                  },
                                                    "security":  [
                                                                     {
                                                                         "bearerAuth":  [

                                                                                        ]
                                                                     }
                                                                 ]
                                                }
                                   },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "UpdateCourseRequest":  {
                                                                   "type":  "object",
                                                                   "properties":  {
                                                                                      "code":  {
                                                                                                   "type":  "string",
                                                                                                   "maxLength":  255,
                                                                                                   "minLength":  0
                                                                                               },
                                                                                      "name":  {
                                                                                                   "type":  "string",
                                                                                                   "maxLength":  255,
                                                                                                   "minLength":  0
                                                                                               },
                                                                                      "studentCount":  {
                                                                                                           "type":  "integer",
                                                                                                           "format":  "int32"
                                                                                                       }
                                                                                  },
                                                                   "required":  [
                                                                                    "code",
                                                                                    "name",
                                                                                    "studentCount"
                                                                                ]
                                                               },
...
                                       "CreateCourseRequest":  {
                                                                   "type":  "object",
                                                                   "properties":  {
                                                                                      "code":  {
                                                                                                   "type":  "string",
                                                                                                   "maxLength":  255,
                                                                                                   "minLength":  0
                                                                                               },
                                                                                      "name":  {
                                                                                                   "type":  "string",
                                                                                                   "maxLength":  255,
                                                                                                   "minLength":  0
                                                                                               },
                                                                                      "studentCount":  {
                                                                                                           "type":  "integer",
                                                                                                           "format":  "int32"
                                                                                                       }
                                                                                  },
                                                                   "required":  [
                                                                                    "code",
                                                                                    "name",
                                                                                    "studentCount"
                                                                                ]
                                                               },
...
                                   },
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}